Refactor import paths and add delay utility function
