--- src/utils/delay.ts
@@ -1,5 +1,0 @@
-export async function delay(time: number) {
-    // console.log('delay');
-    
-    return new Promise((resolve) => setTimeout(resolve, time));
-}
