+++ src/utils/findOrCreateAta.function.ts
@@ -3,7 +3,7 @@ import { Cluster, Connection, PublicKey, TransactionInstruction } from "@solana/
 import { SOLANA_SPL_ATA_PROGRAM_ID, SWAP_PROGRAM_ID, SWAP_PROGRAM_ID_DEV } from "../utils/const";
 import { Program } from "@coral-xyz/anchor";
 import { CreateAssociatedTokenAccountInstructionData } from "./types";
-import { delay } from "./delay";
+import { delay } from "./utils";
 import { getProgram } from "./getProgram.obj";
 
 export async function findOrCreateAta(Data: {
