+++ src/utils/utils.ts
@@ -1,0 +1,17 @@
+import { TxWithSigner, VTxWithSigner} from "./types";
+
+export async function delay(time: number) {
+    // console.log('delay');
+    
+    return new Promise((resolve) => setTimeout(resolve, time));
+}
+
+export function isVersionedTx(tx: TxWithSigner | VTxWithSigner): tx is VTxWithSigner {
+    return (tx as VTxWithSigner).vtx !== undefined;
+}
+
+export function isVersionedArray(txs: TxWithSigner[] | VTxWithSigner[]): txs is VTxWithSigner[] {
+    if (txs.length === 0) return false;
+
+    return isVersionedTx(txs[0]);
+}
